pyproject.toml
[project]
name = "notepad-automation"
version = "1.0.0"
description = "Production-ready Windows desktop automation using vision-based grounding"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Desktop Automation Engineer" }
]
keywords = ["automation", "vision", "CLIP", "desktop", "notepad"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "Pillow>=10.0.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "pywin32>=306",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
notepad-auto = "notepad_automation.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notepad_automation"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
